Move send reset letter method to AuthController #187
