+++ Cooper.Controllers/AuthController.cs
@@ -5,6 +5,7 @@ using Cooper.Services.Authorization;
 using Cooper.Services.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 
 namespace Cooper.Controllers
 {
@@ -15,13 +16,15 @@ namespace Cooper.Controllers
         private readonly UserRepository userRepository;
         private readonly IConfigProvider configProvider;
         private readonly ISocialAuth socialAuth;
+        private readonly ILogger<AuthController> logger;
 
-        public AuthController(IJwtHandlerService jwtService, ISocialAuth socialAuth, IConfigProvider configProvider)
+        public AuthController(IJwtHandlerService jwtService, ISocialAuth socialAuth, IConfigProvider configProvider, ILogger<AuthController> logger)
         {
             userRepository = new UserRepository(jwtService, configProvider);
 
             this.configProvider = configProvider;
             this.socialAuth = socialAuth;
+            this.logger = logger;
         }
 
         /// <summary>
@@ -78,6 +81,45 @@ namespace Cooper.Controllers
                         result = Ok(new { Token = new TokenFactory(login, configProvider).GetTokenString() });
                     }
                 }
+            }
+
+            return result;
+        }
+
+        /// <summary>
+        /// Send letter to reset password.
+        /// </summary>
+        /// <param name="email">User email</param>
+        /// <response code="200">If letter was sent</response>
+        /// <response code="400">If letter was not sent</response>
+        [HttpPost]
+        [Route("reset/send")]
+        [Consumes("application/json")]
+        [Produces("application/json")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public IActionResult SendLetter(TargetEmail email)
+        {
+            IActionResult result = BadRequest();
+
+            if (email != null)
+            {
+                email.Email = DbTools.SanitizeString(email.Email);
+
+                if (userRepository.IfEmailExists(email.Email))
+                {
+                    // TODO: SENT EMAIL
+                    logger.LogInformation("Password reset email was sent for email {0}.", email.Email);
+                    result = Ok();
+                }
+                else
+                {
+                    logger.LogWarning("Password reset email was not sent, because the passed email does not exist.");
+                }
+            }
+            else
+            {
+                logger.LogWarning("Password reset email was not sent, because the passed email was empty.");
             }
 
             return result;
